requirements.txt
3==0.0.0
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-admin==2.0.1
django-admin-material-dashboard==1.0.10
django-countries==5.3.3
django-crispy-forms==1.7.2
django-environ==0.10.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-filter==22.1
django-modelcluster==6.0
django-multiselectfield==0.1.12
django-permissionedforms==0.1
django-phonenumber-field==7.1.0
django-six==1.0.4
django-taggit==3.1.0
django-treebeard==4.7
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.11.0
filetype==1.2.0
fonttools==4.39.4
freetype-py==2.3.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
l18n==2021.3
matplotlib==3.7.1
mysql-connector-python==8.0.33
numpy==1.24.0
openpyxl==3.1.2
packaging==23.1
pep8==1.7.1
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.2.0
pycairo==1.23.0
pycodestyle==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
screen==1.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
tzdata==2022.5
urllib3==1.24.2
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
wagtail==5.0
webencodings==0.5.1
whitenoise==6.4.0
Willow==1.5
xlwt==1.3.0
